@@ -1,9 +1,7 @@
 import React from 'react';
 import './App.css';
-import { ThemeProvider } from '@material-ui/core/styles';
-import { createMuiTheme } from '@material-ui/core/styles';
-import red from '@material-ui/core/colors/red';
-import teal from '@material-ui/core/colors/teal';
+import { ThemeProvider, createTheme } from '@mui/material/styles';
+import {teal, red} from '@mui/material/colors';
 import ImagesGrid from './components/ImagesGrid';
 import AboutContainer from './components/AboutContainer';
 import TopBar from './components/TopBar';
@@ -12,7 +10,7 @@ import Footer from './components/Footer';
 import SkillsContainer from './components/SkillsContainer'
 import InformationContainer from './components/InformationContainer'
 
-const theme = createMuiTheme({
+const theme = createTheme({
   palette: {
     primary: {
       main: red[500],
