@@ -1,7 +1,7 @@
 import React from 'react';
-import Typography from '@material-ui/core/Typography';
-import Container from '@material-ui/core/Container';
-import FavoriteIcon from '@material-ui/icons/Favorite';
+import Typography from '@mui/material/Typography';
+import Container from '@mui/material/Container';
+import FavoriteIcon from '@mui/icons-material/Favorite';
 
 export default function Footer() {
   return (
@@ -18,7 +18,7 @@ export default function Footer() {
     >
       Made with{' '}
       <FavoriteIcon color="error" style={{ margin: '-2px 5px' }} />
-      in Smyrna, GA.
+      in Woodstock, GA.
     </Typography>
   </Container>
   );
